Add the first menus
I added two new functions. The first is the menu() function which is useful to create any type of menu. The second is an application of this menu to create the main menu. I've also added a switch to choose the right option. Finally, i've created a library where i'll put all the different menus of the game.