Changes on menus
Few changes on the functionmenu() and main_menu() : now the first element of the array should be the title of the menu.
Add a new function : bid_menu. In this function it will display the menu to make a bid : The classic bid when the user enter the bid he want to make and the special ones like the capot, coinche, etc.