Adds on the main
I've added some things in the main to establish the turn for all the players when they must bet. I've also added few functions in Cartes.h to help me to do it. I put a funciton that sort the players in the players to know which player must to play in which order. FInally I've commented my functions to be sure the way it works is clear and to prepare the documentation with doxygen later.