Correction of some bugs about AI
I've corrected a little bug about the AI libray. Effectively, We were checking the 4th element of the field color of a struct card. By example with "Hea_Ki", i was checking 'i' instead of 'K' and it was the source of the problem. I also changed the value of the contract at the beginning of the game which I set to 79 ( it could be 18, 0 or 25 but i chose this value because it's nearer from 80). The project topic isn't really clear about this point but it seems that the first contract is compulsory 80 but in this case the ai can never make a bet of 80 because the compulsory score to reach is already 80. So I decided that the first score to reach is 79 beacuse it is really rare that the ai can have 4 strong card of the same color in a round (all the possible strong card of 1 color must to be in his hand !).