First implementation of high score
I've started the management of high scores. I've created a file ( to keep the highscore after closure of the program) and a new libray containing all the functions linked with the high scores. For the moment I have only implemented the read of the file and the function which stock in a struct the datas of the file. Tomorrow i will manage the writing function which must to sort correctly the struct before rewriting the file.