Highscore finished
I've finished the highscores. It was easier than I thought but it works very well. I've also set some function to help me and especally a function that free the memory for the HIGHSCORE struct. @karakala0  I've set an example in the main to help you understand the way it works. Also the name entered by the player couldn't be longer than 7 or we'll have some trouble while we're attempting to display the highscore... Ready to merge !