Bug correction on highscores
I had multiple bugs on highscores. First I had a windows sound when I wa running the program but in fact, it was just an escape sequence '\a' that I forgot to erase in a printf. secondly, I had a problem with a realloc but it wasn't working, so I ask my teacher and he said me to put a malloc instead of a realloc and now it works ! (In fact, I was shortening arrays of char and that wasn't a good idea, so now I free the string and then I malloc it with the correct dimension). I haven't any other kind of error for the moment, we will make tests in the week to verify that it works properly).