Bug correction
I improved some bugs correction. Now the bet of ai should work ! Especially, the bug with the bet ai. it was due to the temporary variable that  I used in my Bubble sort. I was freeing it but by doing it, I was also freeing the memory of the first place of my array. That was fucking off the first place in the array contining the best color option.