Test with a game
We finish this game pretty quickly we are missing some elements. By example we don't manage entirely the bets. It seems also that we have some bugs on the values of the cards when they are in trump... But at least it works more or less... The highscore board also works, it's very important ! And we have a beautiful library with doxygen. So evrything is not so messy but we did our best :D We also do not forget that this game works it is beacuse Leonard solved all the problems related with the game right in time. we weren't geniuses of the belote we tried   to make it as best as we could and for me it is the more  important regardless of some bugs.